api/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["HDYS-TTBYS <tatehide151@outlook.jp>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.1"
djangorestframework = "3.11.1"
django-cors-headers = "3.4.0"
djangorestframework-simplejwt = "4.6.0"
djoser = "2.0.3"
Pillow = "^8.1.0"
django-environ = "^0.4.5"
dj-database-url = "^0.5.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
